+++ lib/src/curved_navigation_item_pro.dart
@@ -1,3 +1,5 @@
+// ignore_for_file: comment_references
+
 import 'package:flutter/material.dart';
 
 /// A single item in a [CurvedNavBar].